Conflicts:
	scilab/Makefile.in
	scilab/configure
	scilab/configure.ac
	scilab/contrib/Makefile.in
	scilab/desktop/Makefile.in
	scilab/desktop/images/icons/Makefile.in
	scilab/libs/MALLOC/Makefile.in
	scilab/libs/Makefile.in
	scilab/libs/doublylinkedlist/Makefile.in
	scilab/libs/dynamiclibrary/Makefile.in
	scilab/libs/hashtable/Makefile.in
	scilab/libs/libst/Makefile.in
	scilab/modules/Makefile.in
	scilab/modules/action_binding/Makefile.in
	scilab/modules/api_scilab/Makefile.in
	scilab/modules/arnoldi/Makefile.in
	scilab/modules/atoms/Makefile.in
	scilab/modules/boolean/Makefile.in
	scilab/modules/cacsd/Makefile.in
	scilab/modules/call_scilab/Makefile.in
	scilab/modules/commons/Makefile.in
	scilab/modules/compatibility_functions/Makefile.in
	scilab/modules/completion/Makefile.in
	scilab/modules/console/Makefile.in
	scilab/modules/core/Makefile.in
	scilab/modules/core/src/c/getval.c
	scilab/modules/data_structures/Makefile.in
	scilab/modules/demo_tools/Makefile.in
	scilab/modules/development_tools/Makefile.in
	scilab/modules/development_tools/src/fake/Makefile.in
	scilab/modules/differential_equations/Makefile.in
	scilab/modules/dynamic_link/Makefile.in
	scilab/modules/elementary_functions/Makefile.in
	scilab/modules/external_objects/Makefile.in
	scilab/modules/fftw/Makefile.in
	scilab/modules/fileio/Makefile.am
	scilab/modules/fileio/Makefile.in
	scilab/modules/fileio/fileio.vcxproj
	scilab/modules/fileio/includes/gw_fileio.h
	scilab/modules/fileio/sci_gateway/c/gw_fileio.c
	scilab/modules/functions/Makefile.in
	scilab/modules/genetic_algorithms/Makefile.in
	scilab/modules/graph/Makefile.in
	scilab/modules/graphic_export/Makefile.in
	scilab/modules/graphic_objects/Makefile.in
	scilab/modules/graphic_objects/sci_gateway/c/gw_graphic_objects.c
	scilab/modules/graphic_objects/sci_gateway/cpp/sci_graphicfunction.cpp
	scilab/modules/graphics/Makefile.in
	scilab/modules/gui/Makefile.in
	scilab/modules/hdf5/Makefile.in
	scilab/modules/helptools/Makefile.in
	scilab/modules/history_browser/Makefile.in
	scilab/modules/history_manager/Makefile.in
	scilab/modules/integer/Makefile.in
	scilab/modules/interpolation/Makefile.in
	scilab/modules/intersci/Makefile.in
	scilab/modules/io/Makefile.in
	scilab/modules/javasci/Makefile.in
	scilab/modules/jvm/Makefile.in
	scilab/modules/linear_algebra/Makefile.in
	scilab/modules/localization/Makefile.in
	scilab/modules/m2sci/Makefile.in
	scilab/modules/matio/Makefile.in
	scilab/modules/mexlib/Makefile.in
	scilab/modules/modules_manager/Makefile.in
	scilab/modules/optimization/Makefile.in
	scilab/modules/output_stream/Makefile.in
	scilab/modules/overloading/Makefile.in
	scilab/modules/parallel/Makefile.in
	scilab/modules/parameters/Makefile.in
	scilab/modules/polynomials/Makefile.in
	scilab/modules/prebuildjava/Makefile.in
	scilab/modules/preferences/Makefile.in
	scilab/modules/randlib/Makefile.in
	scilab/modules/renderer/Makefile.in
	scilab/modules/scicos/Makefile.in
	scilab/modules/scicos_blocks/Makefile.in
	scilab/modules/scinotes/Makefile.in
	scilab/modules/signal_processing/Makefile.in
	scilab/modules/simulated_annealing/Makefile.in
	scilab/modules/sound/Makefile.in
	scilab/modules/sparse/Makefile.in
	scilab/modules/special_functions/Makefile.in
	scilab/modules/spreadsheet/Makefile.in
	scilab/modules/statistics/Makefile.in
	scilab/modules/string/Makefile.in
	scilab/modules/symbolic/Makefile.in
	scilab/modules/tclsci/Makefile.in
	scilab/modules/threads/Makefile.in
	scilab/modules/time/Makefile.in
	scilab/modules/types/Makefile.in
	scilab/modules/ui_data/Makefile.in
	scilab/modules/umfpack/Makefile.in
	scilab/modules/windows_tools/Makefile.in
	scilab/modules/windows_tools/includes/gw_windows_tools.h
	scilab/modules/windows_tools/sci_gateway/c/gw_windows_tools.c
	scilab/modules/windows_tools/sci_gateway/c/sci_chartooem.c
	scilab/modules/windows_tools/sci_gateway/c/sci_oemtochar.c
	scilab/modules/windows_tools/src/c/windows_tools.vcxproj.filters
	scilab/modules/xcos/Makefile.in
	scilab/modules/xml/Makefile.in

+++ scilab/modules/fileio/src/cpp/mtell.cpp
@@ -5,7 +5,7 @@
  * This file must be used under the terms of the CeCILL.
  * This source file is licensed as described in the file COPYING, which
  * you should have received as part of this distribution.  The terms
- * are also available at    
+ * are also available at
  * http://www.cecill.info/licences/Licence_CeCILL_V2-en.txt
  *
  */
@@ -24,25 +24,25 @@ extern "C"
 
 /*--------------------------------------------------------------------------*/
 long int mtell(int fd)
-{    
+{
     File* pF        = FileManager::getFile(fd);
     long int offset = -1;
-    
-	if(pF == NULL) 
-	{
-		return offset;
-	}
-	
-	#ifdef _MSC_VER
-		#if _WIN64 
-			offset = _ftelli64(pF->getFiledesc()) ;
-		#else
-			offset = ftell(pF->getFiledesc()) ;
-		#endif
-	#else
-	    offset = ftell(pF->getFiledesc()) ;
-	#endif
-	
-	return offset;
+
+    if (pF == NULL)
+    {
+        return offset;
+    }
+
+#ifdef _MSC_VER
+#if _WIN64
+    offset = (long)_ftelli64(pF->getFiledesc()) ;
+#else
+    offset = ftell(pF->getFiledesc()) ;
+#endif
+#else
+    offset = ftell(pF->getFiledesc()) ;
+#endif
+
+    return offset;
 }
 /*--------------------------------------------------------------------------*/
